refactor: 조회전용, 쓰기전용 모델로 분리
