+++ src/main/java/com/knu/ddip/location/application/scheduler/LocationScheduler.java
@@ -1,6 +1,6 @@
 package com.knu.ddip.location.application.scheduler;
 
-import com.knu.ddip.location.infrastructure.repositoroy.redis.RedisLocationExpiryCleanup;
+import com.knu.ddip.location.application.service.LocationWriter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
@@ -9,12 +9,12 @@ import org.springframework.stereotype.Component;
 @RequiredArgsConstructor
 public class LocationScheduler {
 
-    private final RedisLocationExpiryCleanup cleanup;
+    private final LocationWriter locationWriter;
 
     @Scheduled(cron = "0 0 * * * *") // 매 정시
     public void cleanupExpiredUserLocations() {
         long now = System.currentTimeMillis();
-        cleanup.cleanupExpiredUserLocations(now);
+        locationWriter.cleanupExpiredUserLocations(now);
     }
 
 }
