+++ src/main/java/com/knu/ddip/location/application/service/LocationReader.java
@@ -1,0 +1,13 @@
+package com.knu.ddip.location.application.service;
+
+import java.util.List;
+
+public interface LocationReader {
+    void validateLocationByCellId(String cellId);
+
+    List<String> findAllLocationsByCellIdIn(List<String> cellIds);
+
+    List<String> findUserIdsByCellIds(List<String> targetCellIds);
+
+    boolean isCellIdNotInTargetArea(String cellId);
+}
