+++ src/test/java/com/knu/ddip/location/application/service/LocationServiceTest.java
@@ -137,8 +137,6 @@ class LocationServiceTest {
     @Test
     void getNeighborRecipientUserIdsTest() {
         // given
-        int level = 17;
-
         UUID myUserId = UUID.randomUUID();
 
         // 일청담
